requirements.txt
fastapi
uvicorn
pymongo
fitz
numpy
PyMuPDF
config
elasticsearch==8.17.1
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
requests==2.32.3
tavily-python==0.5.0